add support for jump_da64 and drop a32_payload
There's a jump_da64 command in the mtkclient project:
https://github.com/bkerler/mtkclient
But it doesn't work.

By reverse-engineering MT7986 BootROM, I found that:
1. after receving the byte indicating a 64-bit payload and
   replying a status, the BootROM tries to recieve one more
   byte, store it into a global variable and reply another
   status.
2. The only other reference to this global variable is a
   check. It checks if the variable equals to 100. If true,
   it sets CPU start address and reset CPU into aarch64 mode,
   otherwise it stops in an infinite loop.

Thanks to an MT7986 BootROM dump by @dangole

With a working jump_da64, we don't need a second aarch32
payload for mode switching. Drop a32_payload and a32_load_addr,
and add --aarch64 to indicate that the payload sent is aarch64
code.

Signed-off-by: Chuanhong Guo <gch981213@gmail.com>